Fixed warning: implicit conversion loses integer precision: 'VkDeviceSize' (aka 'unsigned long long') to 'size_t' (aka 'unsigned int')

+++ src/render/vulkan/SDL_render_vulkan.c
@@ -2899,7 +2899,7 @@ static int VULKAN_UpdateTextureInternal(SDL_Renderer *renderer, VULKAN_Image *ma
             length = pitch;
         }
         for (VkDeviceSize row = h; row--; ) {
-            SDL_memcpy(dst, src, length);
+            SDL_memcpy(dst, src, (size_t)length);
             src += pitch;
             dst += length;
         }
